refactor ukdctl stop version
