+++ ukdctl/ukdctl.go
@@ -1,11 +1,8 @@
 package main
 
 import (
-	"github.com/myechuri/ukd/server/api"
 	"github.com/myechuri/ukd/ukdctl/cmd"
 	"github.com/spf13/cobra"
-	"golang.org/x/net/context"
-	"google.golang.org/grpc"
 	"log"
 )
 
@@ -22,61 +19,15 @@ var (
 		SuggestFor: []string{"ukdctl"},
 	}
 	serverAddress string // used by rootCmd
-	ukName        string // used by stopCmd
 )
-
-func getServerVersion(client api.UkdClient) {
-	versionRequest := &api.VersionRequest{}
-	reply, err := client.GetVersion(context.Background(), versionRequest)
-	if err != nil {
-		log.Fatalf("could not gather grpc server version: %v", err)
-	}
-	log.Printf("Ukd server version: %d.%d", reply.Major, reply.Minor)
-
-}
-
-func stopUK(client api.UkdClient) {
-	stopRequest := &api.StopRequest{
-		Name: "test app",
-	}
-	reply, _ := client.StopUK(context.Background(), stopRequest)
-	log.Printf("Application unikernel stopped: %t, Info: %s",
-		reply.Success, reply.Info)
-}
 
 func main() {
 
 	rootCmd.PersistentFlags().StringVar(&serverAddress, "server-endpoint", defaultServer, "server IP and Port ('ip:port') to connect to")
 
-	// TODO: TLS
-	conn, err := grpc.Dial(serverAddress, grpc.WithInsecure())
-	if err != nil {
-		log.Fatalf("fail to dial: %v", err)
-	}
-	defer conn.Close()
-	client := api.NewUkdClient(conn)
-
-	var versionCmd = &cobra.Command{
-		Use:   "version",
-		Short: "Get server version",
-		Long:  `Get grpc server version`,
-		Run: func(cmd *cobra.Command, args []string) {
-			getServerVersion(client)
-		},
-	}
-	rootCmd.AddCommand(versionCmd)
-
-	rootCmd.AddCommand(cmd.StartCommand(serverAddress))
-
-	var stopCommand = &cobra.Command{
-		Use:   "stopUK",
-		Short: "Stop a Unikernel",
-		Long:  `Stop a unikernel with given name`,
-		Run: func(cmd *cobra.Command, args []string) {
-			stopUK(client)
-		},
-	}
-	rootCmd.AddCommand(stopCommand)
+	rootCmd.AddCommand(cmd.VersionCommand(),
+		cmd.StartCommand(),
+		cmd.StopCommand())
 
 	if err := rootCmd.Execute(); err != nil {
 		log.Fatalf("failed to execute command: %v", err)
